Initial commit (legacy import)
Imported legacy code from my personal finances spreadsheet
on Google Drive and converted it from a [bound] script
to a [standalone] project so that I can manage it via
[@danthareja]'s awesome [node-google-apps-script]
CLI for Google Apps Script projects.

TODO: refine into test-driven design paradigm.
Most of this code is characteristically disorganized
due to its origin as hacky spreadsheet macros.
I'll be refactoring to patterns and migrating it
towards publication as an [add-on], as well as
hooking it up with a hosted continuous integration
service like [Travis CI] (pretending it's Node.js).

[bound]: https://developers.google.com/apps-script/guides/bound#creating_a_bound_script
[standalone]: https://developers.google.com/apps-script/guides/standalone#creating_a_standalone_script
[@danthareja]: https://github.com/danthareja
[node-google-apps-script]: https://github.com/danthareja/node-google-apps-script
[add-on]: https://developers.google.com/apps-script/add-ons/
[Travis CI]: https://docs.travis-ci.com/user/languages/javascript-with-nodejs